Working stock serialization and loading from SO without displaying graph

+++ Assets/Scripts/Stocks/AlphaVantage/AlphaVantageTimeValue.cs
@@ -3,13 +3,13 @@
 public class AlphaVantageTimeValue
 {
     public DateTime Timestamp { get; set; }
-	public decimal Open { get; set; }
-	public decimal High { get; set; }
-	public decimal Low { get; set; }
-	public decimal Close { get; set; }
-	public decimal Volume { get; set; }
+	public double Open { get; set; }
+	public double High { get; set; }
+	public double Low { get; set; }
+	public double Close { get; set; }
+	public double Volume { get; set; }
 
-    public AlphaVantageTimeValue(DateTime timestamp, decimal open, decimal high, decimal low, decimal close, decimal volume)
+    public AlphaVantageTimeValue(DateTime timestamp, double open, double high, double low, double close, double volume)
     {
         Timestamp = timestamp;
         Open = open;
@@ -21,11 +21,11 @@ public class AlphaVantageTimeValue
     public AlphaVantageTimeValue(string[] values)
 	{
         Timestamp = DateTime.Parse(values[0]);
-        Open = decimal.Parse(values[1]);
-        High = decimal.Parse(values[2]);
-        Low = decimal.Parse(values[3]);
-        Close = decimal.Parse(values[4]);
-        Volume = decimal.Parse(values[5]);
+        Open = double.Parse(values[1]);
+        High = double.Parse(values[2]);
+        Low = double.Parse(values[3]);
+        Close = double.Parse(values[4]);
+        Volume = double.Parse(values[5]);
     }
 
     public static AlphaVantageTimeValue FromCSV(string csvLine)
